Update OBJLoader.java
Improved the parsing performance (important mainly for Android). Example model with 32k vertices took initially 71 seconds and loads now in around 15 seconds. I tried to keep as much as possible of the original structure. Main problem on Android is that the Scanner class is horribly slow and the Pattern class as well so I got rid of them and added my own implementation of the split method.