Adding AWT component rendering within app state
Adding AWT component rendering capabilities for any JMonkeyEngine
renderer (works with LWJGL 3.1.x and compatible with other
implementations)

This implementation differs from the previous one as it's relies on
AppState (compatible with all JME application) and as it use only
produces framebuffer.

With this implementation, it is possible to use AWT component ad
rendering target and so to integrate JMonkey rendering within AWT /
Swing application. This capability was not available with LWJGL3 as the
version 3.1.x does not provide anymore AWT link.