1877: Add checkstyle to project
Add checkstyle plugin to common.gradle
Add a checkstyle.xml based on google checkstyle xml with ammendments.
Checkstyle errors do not fail the build, everything is marked as a
warning, can be refined over time as per comments in #1877.

Checkstyle can be run with `./gradlew checkstyleMain` - it runs for all
projects.