Fix logic errors and exception in SpecGloss pipeline
This PR addresses this issue: 
#2386 
And then I also found another critical logic error in the spec gloss pipeline that was likely there even before riccardo adapted the shader to use structs:

It looks like the specular value taken from specularMaps was being totally ignored and not used in the roughness calculation. The code further down in the shader was calculating the roughness solely based on m_Glossiness with no regard for the glossiness texture maps, which is incorrect.

The logic error and exception should all be fixed in this PR now, but it will need testing from a spec gloss model with packed specularGlossinesMap as well as a model with seperate specular and glossiness maps in order to ensure everything works properly now.