Fix non decimal format floats in TangentUtils.glsllib
I also added a line of code that I had previously removed from the tangent calculations that may fix some minor inaccuracies in shading and lighting direction (although it was very minor and barely noticeable).

I previously removed that line of code because it was also flipping the y component of normal map (which is better done by the NORMAL_TYPE param) but then I realized this line of code also flips the x component of the normal map which is likely important for the tangent calculation that follows. So I just updated it to only multiply the x component by -1 and will leave the y component flipping up to NORMAL_TYPE so users can change it as needed.